@@ -2,15 +2,24 @@
   "name": "embarkjs",
   "version": "0.1.0",
   "description": "JavaScript library for easily interacting with web3 technologies",
-  "main": "./dist/index.js",
-  "browser": "./src/browser.js",
+  "main": "dist/node/index.js",
+  "browser": {
+    "./dist/node/index.js": "./dist/browser/browser.js"
+  },
+  "browserslist": [
+    "last 1 version",
+    "not dead",
+    "> 0.2%"
+  ],
   "files": [
     "dist",
     "embark.min.js",
     "src"
   ],
   "scripts": {
-    "babel": "babel --out-dir dist src",
+    "babel": "npm run babel:browser && npm run babel:node",
+    "babel:browser": "BABEL_ENV=browser babel --out-dir dist/browser src",
+    "babel:node": "BABEL_ENV=node babel --out-dir dist src",
     "build": "npm run clean && npm run babel && npm run webpack",
     "clean": "rimraf dist embark.min.js embarkjs-*.tgz package",
     "http-server": "http-server",
@@ -38,19 +47,22 @@
   },
   "homepage": "https://github.com/embark-framework/EmbarkJS#readme",
   "dependencies": {
-    "@babel/runtime": "^7.0.0-beta.52",
+    "@babel/runtime": "^7.0.0-beta.54",
     "async": "^2.0.1"
   },
   "devDependencies": {
-    "@babel/cli": "7.0.0-beta.52",
-    "@babel/core": "7.0.0-beta.52",
-    "@babel/plugin-transform-runtime": "7.0.0-beta.52",
-    "@babel/preset-env": "7.0.0-beta.52",
+    "@babel/cli": "7.0.0-beta.54",
+    "@babel/core": "7.0.0-beta.54",
+    "@babel/plugin-transform-runtime": "7.0.0-beta.54",
+    "@babel/preset-env": "7.0.0-beta.54",
     "ajv": "6.5.2",
     "babel-loader": "8.0.0-beta.4",
     "http-server": "0.11.1",
     "rimraf": "2.6.2",
-    "webpack": "4.15.1",
+    "webpack": "4.16.1",
     "webpack-cli": "3.0.8"
+  },
+  "engines": {
+    "node" : ">=8.11.3"
   }
 }
