@@ -1,28 +1,8 @@
 const path = require('path');
 
 const standalone = {
-    entry: path.resolve(__dirname, 'src') + '/browser.js',
+    entry: path.join(__dirname, 'dist/browser', 'browser.js'),
     mode: 'production',
-    module: {
-        rules: [
-            {
-                test: /\.js$/,
-                exclude: /(node_modules|bower_components)/,
-                use: {
-                    loader: 'babel-loader',
-                    options: {
-                        babelrc: false,
-                        plugins: [
-                            "@babel/plugin-transform-runtime"
-                        ],
-                        presets: [
-                            ['@babel/env']
-                        ]
-                    }
-                }
-            }
-        ]
-    },
     // optimization: {
     //     minimize: false
     // },
